cdk/requirements.txt
aws-cdk.aws-certificatemanager==1.85.0
aws-cdk.aws-dynamodb==1.85.0
aws-cdk.aws-iam==1.85.0
aws-cdk.aws-lambda==1.85.0
aws-cdk.aws-lambda-event-sources==1.85.0
aws-cdk.aws-route53==1.85.0
aws-cdk.aws-route53-targets==1.85.0
aws-cdk.aws-s3==1.85.0
aws-cdk.aws-s3-deployment==1.85.0
aws-cdk.aws-sqs==1.85.0
aws-cdk.aws-timestream==1.85.0
aws-cdk.core==1.85.0
